Move interpolation functions into smoothing module
